LPS-84717 Move SortKey to SortQuery
