+++ modules/apps/headless-apio/structured-content/structured-content-apio-api/src/main/java/com/liferay/structured/content/apio/architect/sort/Sort.java
@@ -14,8 +14,6 @@
 
 package com.liferay.structured.content.apio.architect.sort;
 
-import java.io.Serializable;
-
 /**
  * Models a Sort param for sorting structured content by different fields and
  * sort directives.
@@ -43,50 +41,6 @@ public class Sort {
 	 */
 	public SortQuery getSortQuery() {
 		return _sortQuery;
-	}
-
-	/**
-	 * Models a Sort Key.
-	 *
-	 * @review
-	 */
-	public static class SortKey implements Serializable {
-
-		/**
-		 * Creates a new sort key.
-		 *
-		 * @param  fieldName - the name of the field
-		 * @param  asc - if the sort should be ascending
-		 * @review
-		 */
-		public SortKey(String fieldName, boolean asc) {
-			_fieldName = fieldName;
-			_asc = asc;
-		}
-
-		/**
-		 * Returns the name of the field.
-		 *
-		 * @return - the name of the field
-		 * @review
-		 */
-		public String getFieldName() {
-			return _fieldName;
-		}
-
-		/**
-		 * Returns if the sort key is ascending or not.
-		 *
-		 * @return - if the sort key is ascending or not
-		 * @review
-		 */
-		public boolean isAscending() {
-			return _asc;
-		}
-
-		private final boolean _asc;
-		private final String _fieldName;
-
 	}
 
 	private final SortQuery _sortQuery;
